requirements.txt
fastapi[standard]
uvicorn
db-sqlite3
pydantic
sqlalchemy
databases
python-dotenv
authlib
itsdangerous